refactor(capabilities): make engine own its kinds and split server and proxy
Move the engine cap's internal control-plane kinds into a per-cap
`engine/kinds.rs` (ADR-0121) and decompose the thick `server.rs` /
`proxy.rs` into cohesion-split directory modules — no behavior change.

Kinds moved (consumed only inside aether-capabilities, embedded in no
staying kind): ForwardEnvelope, RouteEnvelope, CallSettled,
EngineHeartbeatTick, EngineDied, EngineAlive. EngineDied re-imports
DeathReason from aether_kinds (capabilities -> kinds, the allowed
direction). The request/result/descriptor family stays in aether-kinds:
it is the MCP harness's RPC protocol and aether-mcp is barred from
depending on aether-capabilities. The descriptor inventory is global, so
describe_kinds still surfaces every engine kind.

server.rs -> server/{mod,config,artifacts,fleet}.rs and proxy.rs ->
proxy/{mod,config,connect,heartbeat,sinks}.rs. The #[bridge]/#[actor]
macro requires a single inline module with one impl block, so each
actor's handlers and any field-accessing inherent methods stay in mod.rs;
the field-independent config, store/artifact helpers, fleet helpers,
connect/heartbeat sidecars, and test sinks move to sibling modules.
Moving the helpers out of the macro expansion exposed a previously
macro-suppressed doc_lazy_continuation lint, fixed by reflowing one doc.
